featureWriter: import Kern- and MarkFeatureWriter from top
